feat: setup blob sink in e2e tests
