+++ yarn-project/blob-sink/src/config.ts
@@ -1,4 +1,7 @@
+import { DataStoreConfig } from "@aztec/kv-store/config";
+
 export interface BlobSinkConfig {
   port?: number;
+  dataStoreConfig?: DataStoreConfig;
   otelMetricsCollectorUrl?: string;
 }
