+++ yarn-project/blob-sink/src/factory.ts
@@ -1,0 +1,25 @@
+import { BlobSinkService } from "./index.js";
+import { createStore } from "@aztec/kv-store/utils";
+import { BlobSinkConfig } from "./config.js";
+import { AztecKVStore } from "@aztec/kv-store";
+
+
+// If data store settings are provided, the store is created and returned.
+// Otherwise, undefined is returned and an in memory store will be used.
+async function getDataStoreConfig(config?: BlobSinkConfig): Promise<AztecKVStore | undefined> {
+  if (!config?.dataStoreConfig) {
+    return undefined;
+  }
+  return await createStore('blob-sink', config.dataStoreConfig);
+}
+
+// TOOD: telemetry client config too
+
+/**
+ * Creates a blob sink service from the provided config.
+ */
+export async function createBlobSinkService(config?: BlobSinkConfig): Promise<BlobSinkService> {
+  const store = await getDataStoreConfig(config);
+
+  return new BlobSinkService(config, store);
+}
